[STYLE] 멤버관리 페이지 컴포넌트화
